% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enrichment_plot}
\alias{enrichment_plot}
\title{Produce scatterplot}
\usage{
enrichment_plot(data, xvar)
}
\arguments{
\item{df}{(data.frame) table with fedup results}

\item{xVar}{(char) x-axis variable (must be a column value in \code{df})}

\item{yVar}{(char) y-axis variable (must be a column value in \code{df})}

\item{xLab}{(char) x-axis label (default \code{xVar} value)}

\item{yLab}{(char) y-axis label (default NULL)}

\item{pTitle}{(char) plot title (default NULL)}

\item{fillVar}{(char) point fill variable (default NULL)}

\item{fillCol}{(char) point fill colours (default NULL)}

\item{fillLab}{(char) point fill label (default \code{fillVar} value)}

\item{sizeVar}{(char) point size variable (default NULL)}

\item{sizeLab}{(char) point size label (default \code{sizeVar} value)}
}
\value{
Object returned from ggplot with the enrichment dot plot
}
\description{
Produce scatterplot
}
